move inline content to utils
