+++ wandb/apis/reports/util.py
@@ -364,3 +364,41 @@ def shift(p1: Panel, p2: Panel) -> Tuple[Panel, Panel]:
     y = l1["y"] + l1["h"] - l2["y"]
 
     return x, y
+
+
+class InlineLaTeX(Base):
+    latex: str = Attr()
+
+    def __init__(self, latex="", *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.latex = latex
+
+    @property
+    def spec(self) -> dict:
+        return {"type": "latex", "children": [{"text": ""}], "content": self.latex}
+
+
+class InlineCode(Base):
+    code: str = Attr()
+
+    def __init__(self, code="", *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.code = code
+
+    @property
+    def spec(self) -> dict:
+        return {"text": self.code, "inlineCode": True}
+
+
+class Link(Base):
+    text: str = Attr()
+    url: str = Attr()
+
+    def __init__(self, text, url, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.text = text
+        self.url = url
+
+    @property
+    def spec(self) -> dict:
+        return {"type": "link", "url": self.url, "children": [{"text": self.text}]}
